+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # :package: 0.1.0
 
 01. :gift: Added interface `dev.orne.test.rnd.Generator`
+01. :gift: Added class `dev.orne.test.rnd.AbstractGenerator`